pyproject.toml
[project]
name = "ImageHash"
version = "4.3.2"
description = "Image Hashing library"
readme = "README.md"
authors = [
    { name = "Johannes Buchner", email = "buchner.johannes@gmx.at" }
]
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "scipy",
    "pillow",
    "PyWavelets",
    "rust-just>=1.41.0",
]
license = { text = "2-clause BSD License" }
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
]

[project.urls]
Homepage = "https://github.com/JohannesBuchner/imagehash"
Repository = "https://github.com/JohannesBuchner/imagehash"

[project.scripts]
find_similar_images = "find_similar_images:main"

[project.optional-dependencies]
test = ["pytest>=3"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/imagehash"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/find_similar_images.py",
    "/README.md",
    "/LICENSE",
]

[dependency-groups]
dev = [
    "coverage>=7.2.7",
    "mypy>=1.4.1",
    "pytest>=7.4.4",
    "ruff>=0.12.2",
    "scipy-stubs>=1.15.3.0",
    "six>=1.17.0",
    "types-six>=1.17.0.20241205",
]

[tool.ruff]
target-version = "py310"
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
		"W191"
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
"tests/**/*" = ["B011"]   # assert False in tests

[tool.ruff.format]
quote-style = "single"
indent-style = "tab"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
explicit_package_bases = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
